+++ app/controllers/api/v1/book_suggestions_controller.rb
@@ -3,7 +3,7 @@ module Api
     class BookSuggestionsController < ApiController
       skip_before_action :authenticate_user!, only: :create
       def create
-        book_suggestion = BookSuggestionService.new(book_suggestion_params, current_user).new_book_suggestion
+        book_suggestion = Api::V1::BookSuggestionService.new(book_suggestion_params, current_user).new_book_suggestion
         if book_suggestion.save
           render json: book_suggestion, status: :created
         else
